fetch functions refactoring
